requirements.txt
python-binance==1.0.19
typer[all]>=0.12.3
python-dotenv==1.0.0
